+++ openbb_sdk/openbb_sdk_core/openbb_sdk_core/app/provider_interface.py
@@ -154,7 +154,6 @@ class ProviderInterface:
         cls,
         providers: Any,
     ) -> Tuple[Dict[str, Tuple[str, Any, Any]], Dict[str, Tuple[str, Any, Any]]]:
-
         standard: Dict[str, Tuple[str, Any, Any]] = {}
         extra: Dict[str, Tuple[str, Any, Any]] = {}
 
@@ -243,7 +242,6 @@ class ProviderInterface:
         cls,
         providers: Any,
     ) -> Tuple[Dict[str, Tuple[str, Any, Any]], Dict[str, Tuple[str, Any, Any]]]:
-
         standard: Dict[str, Tuple[str, Any, Any]] = {}
         extra: Dict[str, Tuple[str, Any, Any]] = {}
